refactor(Sources): Move sources into `cloud` crate
