--- rust/buildpack-sources/Cargo.toml
@@ -1,19 +1,0 @@
-[package]
-name = "buildpack-sources"
-description = "A Cloud Native Buildpack for importing project sources"
-version = "0.0.0"
-edition = "2021"
-
-[features]
-# Only use sources that are likely to be useful for building images
-# (e.g. ignore `doi` and others for now)
-default = [
-    "sources/gdrive",
-    "sources/github",
-    "sources/gitlab",
-    "sources/http",
-]
-
-[dependencies]
-sources = { path = "../sources", default-features = false }
-buildpack = { path = "../buildpack" }
