--- rust/buildpack-sources/buildpack.toml
@@ -1,17 +1,0 @@
-api = "0.6"
-
-[buildpack]
-id = "stencila/sources"
-version = "0.1.0"
-name = "Stencila Sources Buildpack"
-keywords = ["stencila"]
-homepage = "https://github.com/stencila/stencila/tree/HEAD/rust/buildpack-sources"
-
-[[buildpack.licenses]]
-type = "Apache-2.0"
-
-[metadata]
-creates-layer = false
-
-[[stacks]]
-id = "io.buildpacks.stacks.bionic"
