--- rust/buildpack-sources/src/main.rs
@@ -1,4 +1,0 @@
-use buildpack::buildpack_main;
-use buildpack_sources::SourcesBuildpack;
-
-buildpack_main!(SourcesBuildpack);
