+++ rust/cloud/src/lib.rs
@@ -5,6 +5,9 @@ mod utils;
 pub mod cli;
 pub mod orgs;
 pub mod projects;
+pub mod sources;
 pub mod teams;
 pub mod tokens;
 pub mod users;
+
+pub use types::ProjectLocal;
