+++ rust/cloud/src/projects.rs
@@ -18,66 +18,79 @@ use crate::{
     utils::token_read,
 };
 
-/// Read the project in a directory, if any
-///
-/// Returns the path of the project file and the local representation of the project.
-pub fn project_read(dir: &Path) -> Result<Option<(PathBuf, ProjectLocal)>> {
-    for filename in ["stencila.yaml", "stencila.toml", "stencila.json"] {
-        let path = dir.join(filename);
-        if path.exists() {
-            let content = read_to_string(&path)?;
-            let path_str = path.to_string_lossy();
-            let project = if path_str.ends_with(".yaml") {
-                serde_yaml::from_str(&content)?
-            } else if path_str.ends_with(".toml") {
-                toml::from_str(&content)?
-            } else if path_str.ends_with(".json") {
-                serde_json::from_str(&content)?
-            } else {
-                unreachable!()
-            };
-            return Ok(Some((path, project)));
-        }
-    }
-    Ok(None)
-}
-
-/// Write the project back to a path
-pub fn project_write(path: &Path, project: ProjectLocal) -> Result<()> {
-    let path_str = path.to_string_lossy();
-    let contents = if path_str.ends_with(".yaml") {
-        serde_yaml::to_string(&project)?
-    } else if path_str.ends_with(".toml") {
-        toml::to_string_pretty(&project)?
-    } else if path_str.ends_with(".json") {
-        serde_json::to_string_pretty(&project)?
-    } else {
-        bail!("Unsupported extension for project file")
-    };
-    write(path, contents)?;
-    Ok(())
-}
-
-/// Resolve the current project
-///
-/// Walks up the directory tree until a project file e.g. `stencila.yaml` is found.
-/// Returns the path of the project file and the local representation of the project.
-pub fn project_current() -> Result<(PathBuf, ProjectLocal)> {
-    let mut dir = env::current_dir()?;
-    loop {
-        if let Some(pair) = project_read(&dir)? {
-            return Ok(pair);
-        }
-        if let Some(parent) = dir.parent() {
-            dir = parent.to_path_buf();
-        } else {
-            break;
-        }
-    }
-    bail!("Unable to determine the current project. Could not find a `stencila.yaml`, `stencila.toml` or `stencila.json` file in the current directory or any of its parents")
-}
-
 impl ProjectLocal {
+    /// Resolve the current project
+    ///
+    /// Walks up the directory tree until a project file e.g. `stencila.yaml` is found.
+    /// Returns the path of the project file and the local representation of the project.
+    pub fn current() -> Result<Self> {
+        let mut dir = env::current_dir()?;
+        loop {
+            if let Some(project) = Self::read(&dir)? {
+                return Ok(project);
+            }
+            if let Some(parent) = dir.parent() {
+                dir = parent.to_path_buf();
+            } else {
+                break;
+            }
+        }
+        bail!("Unable to determine the current project. Could not find a `stencila.yaml`, `stencila.toml` or `stencila.json` file in the current directory or any of its parents")
+    }
+
+    /// Read the project in a directory, if any
+    ///
+    /// Returns the path of the project file and the local representation of the project.
+    pub fn read(dir: &Path) -> Result<Option<Self>> {
+        for filename in ["stencila.yaml", "stencila.toml", "stencila.json"] {
+            let path = dir.join(filename);
+            if path.exists() {
+                let content = read_to_string(&path)?;
+                let path_str = path.to_string_lossy();
+                let mut project: ProjectLocal = if path_str.ends_with(".yaml") {
+                    serde_yaml::from_str(&content)?
+                } else if path_str.ends_with(".toml") {
+                    toml::from_str(&content)?
+                } else if path_str.ends_with(".json") {
+                    serde_json::from_str(&content)?
+                } else {
+                    unreachable!()
+                };
+                project.path = path;
+                return Ok(Some(project));
+            }
+        }
+        Ok(None)
+    }
+
+    /// Get the path of a project file
+    pub fn path(&self) -> &Path {
+        self.path.as_path()
+    }
+
+    /// Get the directory of a project
+    pub fn dir(&self) -> &Path {
+        self.path
+            .parent()
+            .expect("Project file should always have a parent")
+    }
+
+    /// Write the project back to a path
+    pub fn write(&self) -> Result<()> {
+        let path_str = self.path.to_string_lossy();
+        let contents = if path_str.ends_with(".yaml") {
+            serde_yaml::to_string(&self)?
+        } else if path_str.ends_with(".toml") {
+            toml::to_string_pretty(&self)?
+        } else if path_str.ends_with(".json") {
+            serde_json::to_string_pretty(&self)?
+        } else {
+            bail!("Unsupported extension for project file: {}", path_str)
+        };
+        write(&self.path, contents)?;
+        Ok(())
+    }
+
     /// Convert a local project to a JSON payload suitable for a `PATCH /projects/<id>` request
     fn to_json(&self) -> Result<serde_json::Value> {
         Ok(json!({
@@ -145,7 +158,7 @@ pub async fn project_create(
         return Error::from_response(response).await;
     };
 
-    if project_current().is_err() {
+    if ProjectLocal::current().is_err() {
         let yaml = serde_yaml::to_string(&project)?;
         write("stencila.yaml", yaml)?;
     }
@@ -204,7 +217,7 @@ pub async fn project_retrieve(project_id: &str) -> Result<ProjectRemote> {
 }
 
 pub async fn project_pull() -> Result<()> {
-    let (path, project) = project_current()?;
+    let project = ProjectLocal::current()?;
 
     let id = match project.id {
         Some(id) => id,
@@ -219,13 +232,15 @@ pub async fn project_pull() -> Result<()> {
         .send()
         .await?;
     if response.status().is_success() {
-        let project: ProjectLocal = response.json().await?;
-        project_write(&path, project)?;
+        let path = project.path();
+        let mut project: ProjectLocal = response.json().await?;
+        project.path = path.to_path_buf();
+        project.write()?;
 
         tracing::info!(
             "Successfully pulled project #{} from Stencila Cloud to file `{}`",
             id,
-            path.display()
+            project.path().display()
         );
 
         Ok(())
@@ -235,7 +250,7 @@ pub async fn project_pull() -> Result<()> {
 }
 
 pub async fn project_push() -> Result<()> {
-    let (path, project) = project_current()?;
+    let project = ProjectLocal::current()?;
 
     let id = match project.id {
         Some(id) => id,
@@ -256,7 +271,7 @@ pub async fn project_push() -> Result<()> {
         tracing::info!(
             "Successfully pushed project  #{} from file `{}` to Stencila Cloud",
             id,
-            path.display()
+            project.path().display()
         );
 
         Ok(())
@@ -275,10 +290,10 @@ pub async fn project_delete(project_id: &str) -> Result<()> {
         return Error::from_response(response).await;
     }
 
-    if let Ok((path, project)) = project_current() {
+    if let Ok(project) = ProjectLocal::current() {
         if let Some(id) = project.id {
             if id.to_string() == project_id {
-                remove_file(path)?;
+                remove_file(project.path())?;
             }
         }
     }
@@ -422,7 +437,7 @@ pub mod cli {
         pub fn resolve(&self) -> eyre::Result<String> {
             match &self.project {
                 Some(id) => Ok(id.to_string()),
-                None => match project_current()?.1.id {
+                None => match ProjectLocal::current()?.id {
                     Some(id) => Ok(id.to_string()),
                     None => bail!("Stencila project file does not have an id"),
                 },
