--- rust/sources/Cargo.toml
@@ -1,21 +1,0 @@
-[package]
-edition = "2021"
-name = "sources"
-version = "0.0.0"
-
-[features]
-doi = ["providers/provider-doi"]
-elife = ["providers/provider-elife"]
-gdrive = ["providers/provider-gdrive"]
-github = ["providers/provider-github"]
-gitlab = ["providers/provider-gitlab"]
-http = ["providers/provider-http"]
-default = ["doi", "elife", "gdrive", "github", "gitlab", "http"]
-
-[dependencies]
-common = { path = "../common" }
-cron-utils = { path = "../cron-utils" }
-files = { path = "../files" }
-graph-triples = { path = "../graph-triples" }
-providers = { path = "../providers" }
-stencila-schema = { path = "../schema" }
