+++ rust/stencila/Cargo.toml
@@ -20,7 +20,6 @@ parsers-cli = ["parsers/cli"]
 kernels-cli = ["kernels/cli"]
 binaries-cli = ["binaries/cli"]
 providers-cli = ["providers/cli"]
-buildpacks-cli = ["buildpacks/cli"]
 images-cli = ["images/cli"]
 
 server = [
@@ -111,16 +110,6 @@ default = [
   "providers/provider-gitlab",
   "providers/provider-http",
 
-  # Buildpacks
-  "buildpacks-cli",
-  "buildpacks/buildpack-apt",
-  "buildpacks/buildpack-dockerfile",
-  "buildpacks/buildpack-node",
-  "buildpacks/buildpack-python",
-  "buildpacks/buildpack-r",
-  "buildpacks/buildpack-sources",
-  "buildpacks/buildpack-stencila",
-
   # Images
   "images-cli",
 
@@ -150,7 +139,6 @@ validator = { version = "0.15.0", features = ["derive"] }
 # Required internal crates
 
 binaries = { path = "../binaries" }
-buildpacks = { path = "../buildpacks" }
 cloud = { path = "../cloud" }
 codecs = { path = "../codecs" }
 common = { path = "../common" }
@@ -173,7 +161,6 @@ node-reshape = { path = "../node-reshape" }
 parsers = { path = "../parsers" }
 path-utils = { path = "../path-utils" }
 providers = { path = "../providers" }
-sources = { path = "../sources" }
 tasks = { path = "../tasks" }
 utils = { path = "../utils" }
 uuids = { path = "../uuids" }
