--- rust/tasks/taskfiles/sources.yaml
@@ -1,9 +1,0 @@
-version: '3'
-
-tasks:
-  pull:
-    desc: Pull all the sources of the current project
-    summary: |
-      Use this task to pull each of the sources listed in the project configuration file (e.g. `stencila.yaml`).
-    cmds:
-      - stencila sources pull
